Changing the peak finding.
Changing the way peaks are identified by using the scipy find_peaks function. This function allows for more or other conditions to be used when identifying peaks in a signal. 

Also adding an option to weigh each event equally in addition to the already existing amplitude weighting when calculating the average.